fix(avatar)!: remove BaseAvatar
Closes #2617
Closes #2575
Closes #2349
