+++ .changeset/remove-baseavatar.md
@@ -1,0 +1,4 @@
+---
+"@patternfly/elements": major
+---
+`<pf-avatar>`: Removed `BaseAvatar` class. Reimplement (recommended) or extend `PfAvatar`.
