--- elements/pf-avatar/BaseAvatar.css
@@ -1,13 +1,0 @@
-:host {
-  display: contents;
-}
-
-svg,
-:host([src]) img {
-  display: inline;
-  object-fit: cover;
-}
-
-:host([hidden]) {
-  display: none;
-}
